+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and start a new "In Progress" section above it.
 ## In progress: 0.135.0
 
 - Have `integrations.s3` for interaction with Object Storage that follows the S3 API.
+- `ElasticJobRegistry`: add support for pre-serialization of process graph ([Open-EO/openeo-geopyspark-driver#1232](https://github.com/Open-EO/openeo-geopyspark-driver/issues/1232))
 
 ## 0.134.0